MANUS_fractal_app/requirements.txt
numpy>=1.20.0
scipy>=1.6.0
matplotlib>=3.3.0
pillow>=8.0.0